+++ isy-ueberwachung/src/main/java/de/bund/bva/isyfact/ueberwachung/autoconfigure/IsyUeberwachungAutoConfiguration.java
@@ -5,15 +5,15 @@ import java.util.Map;
 
 import javax.servlet.ServletRegistration;
 
-import de.bund.bva.isyfact.ueberwachung.config.LoadbalancerServletConfigurationProperties;
-import de.bund.bva.isyfact.ueberwachung.config.NachbarsystemConfigurationProperties;
-import de.bund.bva.isyfact.ueberwachung.service.loadbalancer.LoadbalancerServlet;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import de.bund.bva.isyfact.ueberwachung.config.LoadbalancerServletConfigurationProperties;
+import de.bund.bva.isyfact.ueberwachung.service.loadbalancer.LoadbalancerServlet;
 
 @Configuration
 @EnableConfigurationProperties