+++ isy-ueberwachung/src/test/java/de/bund/bva/isyfact/ueberwachung/actuate/health/IsyHealthEndpointTest.java
@@ -1,13 +1,14 @@
 package de.bund.bva.isyfact.ueberwachung.actuate.health;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.mockito.Mockito.mock;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.HealthIndicator;
 import org.springframework.boot.actuate.health.Status;
-
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
 
 /**
  * Test für den initialen Zustand des Endpoints